requirements.txt
Django==3.0.6
requests==2.23.0
mysqlclient==1.4.6
django-webpack-loader==0.7.0
pytz==2018.7
djangorestframework==3.11.0
django-cors-headers==3.2.1
